moved some classes to Model.Utils deleted useless class(move) and methods
