+++ quoridorcore/src/main/java/com/harush/zitoon/quoridor/core/Utils/PopulateBoardUtil.java
@@ -1,4 +1,6 @@
-package com.harush.zitoon.quoridor.core.model;
+package com.harush.zitoon.quoridor.core.Utils;
+
+import com.harush.zitoon.quoridor.core.model.Board;
 
 import java.util.List;
