+++ quoridorcore/src/main/java/com/harush/zitoon/quoridor/core/Utils/PopulateBoardUtilImpl.java
@@ -1,4 +1,7 @@
-package com.harush.zitoon.quoridor.core.model;
+package com.harush.zitoon.quoridor.core.Utils;
+
+import com.harush.zitoon.quoridor.core.model.Board;
+import com.harush.zitoon.quoridor.core.model.WallData;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -34,7 +37,7 @@ public class PopulateBoardUtilImpl implements PopulateBoardUtil {
     @Override
     public void populateBoardWithWalls(Board board, List<PlayerHistory> playerHistories) {
         List<WallData> wallData = playerHistories.stream().
-                filter(playerHistory -> playerHistory.getWallPlacements() != null).
+                filter(playerHistory -> playerHistory.getWallPlacements() != null).     //null? not -1?
                 flatMap(playerHistory -> playerHistory.getWallPlacements().stream())
                 .collect(Collectors.toList());
