+++ quoridorcore/src/main/java/com/harush/zitoon/quoridor/core/Utils/WinnerDecider.java
@@ -1,0 +1,8 @@
+package com.harush.zitoon.quoridor.core.Utils;
+
+import com.harush.zitoon.quoridor.core.model.Player;
+
+public interface WinnerDecider {
+
+    boolean isWinner(Player player);
+}
