+++ quoridorcore/src/main/java/com/harush/zitoon/quoridor/core/Utils/WinnerDeciderLogic.java
@@ -1,4 +1,7 @@
-package com.harush.zitoon.quoridor.core.model;
+package com.harush.zitoon.quoridor.core.Utils;
+
+import com.harush.zitoon.quoridor.core.model.Coordinate;
+import com.harush.zitoon.quoridor.core.model.Player;
 
 import java.util.ArrayList;
 import java.util.HashMap;
