+++ quoridorcore/src/main/java/com/harush/zitoon/quoridor/core/dao/dbo/converter/PlayerAction2GameRecDBOConverter.java
@@ -1,7 +1,7 @@
 package com.harush.zitoon.quoridor.core.dao.dbo.converter;
 
 import com.harush.zitoon.quoridor.core.dao.dbo.GameRecDBO;
-import com.harush.zitoon.quoridor.core.model.PlayerAction;
+import com.harush.zitoon.quoridor.core.Utils.PlayerAction;
 
 public interface PlayerAction2GameRecDBOConverter {
