+++ quoridorcore/src/main/java/com/harush/zitoon/quoridor/core/model/VerticalWallLogic.java
@@ -1,5 +1,8 @@
 package com.harush.zitoon.quoridor.core.model;
 
+import com.harush.zitoon.quoridor.core.Utils.PlayerAction;
+import com.harush.zitoon.quoridor.core.Utils.Settings;
+
 public class VerticalWallLogic implements Wall {
 
     private GameSession gameSession;
