--- quoridorcore/src/main/java/com/harush/zitoon/quoridor/core/model/WinnerDecider.java
@@ -1,6 +1,0 @@
-package com.harush.zitoon.quoridor.core.model;
-
-public interface WinnerDecider {
-
-    boolean isWinner(Player player);
-}
