+++ quoridorcore/src/test/java/com/harush/zitoon/quoridor/core/model/GamePersistenceServiceTest.java
@@ -1,20 +1,15 @@
 package com.harush.zitoon.quoridor.core.model;
 
-import com.google.common.collect.Lists;
-import com.google.common.collect.Sets;
+import com.harush.zitoon.quoridor.core.Utils.GamePersistenceServiceImpl;
+import com.harush.zitoon.quoridor.core.Utils.PlayersHistoryFactory;
+import com.harush.zitoon.quoridor.core.Utils.PopulateBoardUtil;
+import com.harush.zitoon.quoridor.core.Utils.Statistics;
 import com.harush.zitoon.quoridor.core.dao.GameRecDAO;
-import com.harush.zitoon.quoridor.core.dao.dbo.GameRecDBO;
-import org.junit.Assert;
 import org.junit.Before;
-import org.junit.Test;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.MockitoAnnotations;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Set;
 
 import static org.mockito.Matchers.eq;
 import static org.mockito.Mockito.when;
