+++ quoridorcore/src/test/java/com/harush/zitoon/quoridor/core/model/PlayerAction2GameRecDBOConverterTest.java
@@ -1,5 +1,6 @@
 package com.harush.zitoon.quoridor.core.model;
 
+import com.harush.zitoon.quoridor.core.Utils.PlayerAction;
 import com.harush.zitoon.quoridor.core.dao.dbo.GameRecDBO;
 import com.harush.zitoon.quoridor.core.dao.dbo.converter.PlayerAction2GameRecDBOConverter;
 import com.harush.zitoon.quoridor.core.dao.dbo.converter.PlayerAction2GameRecDBOConverterImpl;
