+++ quoridorcore/src/test/java/com/harush/zitoon/quoridor/core/model/PlayerHistoryFactoryTest.java
@@ -2,6 +2,8 @@ package com.harush.zitoon.quoridor.core.model;
 
 import com.google.common.collect.Lists;
 import com.google.common.collect.Sets;
+import com.harush.zitoon.quoridor.core.Utils.PlayerHistory;
+import com.harush.zitoon.quoridor.core.Utils.PlayersHistoryFactoryImpl;
 import com.harush.zitoon.quoridor.core.dao.GameRecDAO;
 import com.harush.zitoon.quoridor.core.dao.PlayerDAO;
 import com.harush.zitoon.quoridor.core.dao.dbo.GameRecDBO;
