+++ quoridorui/src/main/java/com/harush/zitoon/quoridor/ui/controller/SetupController.java
@@ -1,5 +1,6 @@
 package com.harush.zitoon.quoridor.ui.controller;
 
+import com.harush.zitoon.quoridor.core.Utils.*;
 import com.harush.zitoon.quoridor.core.dao.*;
 import com.harush.zitoon.quoridor.core.dao.dbo.GameDBO;
 import com.harush.zitoon.quoridor.core.dao.dbo.converter.Player2PlayerDBOConverter;
